XD-1849 Move instances by jobName endpoint
- Move from `/batch/jobs/{jobName}/instances` to `/jobs/instances?jobname={jobName}`
- Move test
