+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/rest/BatchJobInstancesController.java
@@ -17,6 +17,7 @@
 package org.springframework.xd.dirt.rest;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 import org.springframework.batch.core.JobExecution;
@@ -28,6 +29,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.xd.dirt.job.JobExecutionInfo;
@@ -74,6 +76,39 @@ public class BatchJobInstancesController extends AbstractBatchJobsController {
 		}
 		catch (NoSuchJobInstanceException e) {
 			throw new NoSuchBatchJobInstanceException(instanceId);
+		}
+	}
+
+	/**
+	 * Return a paged collection of job instances for a given job.
+	 *
+	 * @param jobName name of the batch job
+	 * @param startJobInstance start index for the job instance
+	 * @param pageSize page size for the list
+	 * @return collection of JobInstances by job name
+	 */
+	@RequestMapping(value = "", method = RequestMethod.GET, params = "jobname")
+	@ResponseStatus(HttpStatus.OK)
+	public Collection<JobInstanceInfoResource> instancesForJob(@RequestParam("jobname") String jobName,
+			@RequestParam(defaultValue = "0") int startJobInstance, @RequestParam(defaultValue = "20") int pageSize) {
+
+		try {
+			Collection<JobInstance> jobInstances = jobService.listJobInstances(jobName, startJobInstance, pageSize);
+			List<JobInstanceInfoResource> result = new ArrayList<JobInstanceInfoResource>();
+			for (JobInstance jobInstance : jobInstances) {
+				List<JobExecution> jobExecutions = (List<JobExecution>) jobService.getJobExecutionsForJobInstance(
+						jobName, jobInstance.getId());
+				List<JobExecutionInfo> jobExecutionInfos = new ArrayList<JobExecutionInfo>();
+				for (JobExecution jobExecution : jobExecutions) {
+					jobExecutionInfos.add(new JobExecutionInfo(jobExecution, timeZone));
+				}
+				result.add(jobInstanceInfoResourceAssembler.toResource(new JobInstanceInfo(jobInstance,
+						jobExecutionInfos)));
+			}
+			return result;
+		}
+		catch (NoSuchJobException e) {
+			throw new NoSuchBatchJobException(jobName);
 		}
 	}
 }
